requirements.txt
playwright==1.49.1
playwright-stealth==2.0.2
gspread==6.1.4
google-auth==2.37.0
python-telegram-bot==21.9
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.5
2captcha-python==1.5.0
